+++ Examples/ZiggeoDemoSwift/ZiggeoDemoSwift/ViewControllers/Home/RecordingDetailViewController.swift
@@ -29,7 +29,7 @@ class RecordingDetailViewController: UIViewController {
     @IBOutlet weak var saveButton: UIView!
     
     // MARK: - Public variables
-    var mediaType = Media_Type.Video
+    var mediaType = VIDEO
     var recording: ContentModel?
     var recordingDelegate: RecordingDelegate?
     
@@ -42,11 +42,11 @@ class RecordingDetailViewController: UIViewController {
         // Do any additional setup after loading the view.
 
         if (recording != nil) {
-            if (mediaType == Media_Type.Video) {
+            if (mediaType == VIDEO) {
                 imageView.contentMode = UIView.ContentMode.scaleAspectFill
                 let imageUrlString = Common.ziggeo?.videos.getImageUrl(recording!.token)
                 imageView.setURL(imageUrlString, placeholder: nil)
-            } else if (mediaType == Media_Type.Audio) {
+            } else if (mediaType == AUDIO) {
                 imageView.contentMode = UIView.ContentMode.scaleAspectFit
                 imageView.image = UIImage(named: "bg_audio")
             } else {
@@ -86,7 +86,7 @@ class RecordingDetailViewController: UIViewController {
         alert.addAction(UIAlertAction(title: "Yes", style: UIAlertAction.Style.default, handler: { [self] (action) in
             if (self.recording != nil) {
                 SVProgressHUD.show()
-                if (self.mediaType == Media_Type.Video) {
+                if (self.mediaType == VIDEO) {
                     Common.ziggeo?.videos.destroy(self.recording!.token, callback: { jsonObject, response, error in
                         SVProgressHUD.dismiss()
                         self.isEditMode = false
@@ -94,7 +94,7 @@ class RecordingDetailViewController: UIViewController {
                         self.navigationController?.popViewController(animated: true)
                     })
                     
-                } else if (self.mediaType == Media_Type.Audio) {
+                } else if (self.mediaType == AUDIO) {
                     Common.ziggeo?.audios.destroy(self.recording!.token, callback: { jsonObject, response, error in
                         SVProgressHUD.dismiss()
                         self.isEditMode = false
@@ -121,7 +121,7 @@ class RecordingDetailViewController: UIViewController {
             let data = ["title": titleTextField.text,
                         "description": descriptionTextField.text]
             
-            if (self.mediaType == Media_Type.Video) {
+            if (self.mediaType == VIDEO) {
                 Common.ziggeo?.videos.update(recording!.token, data: data as [AnyHashable : Any], callback: { content, response, error in
                     SVProgressHUD.dismiss()
                     
@@ -129,7 +129,7 @@ class RecordingDetailViewController: UIViewController {
                     self.refreshButtons()
                 })
                 
-            } else if (self.mediaType == Media_Type.Audio) {
+            } else if (self.mediaType == AUDIO) {
                 Common.ziggeo?.audios.update(recording!.token, data: data as [AnyHashable : Any], callback: { content, response, error in
                     SVProgressHUD.dismiss()
                     
@@ -150,7 +150,7 @@ class RecordingDetailViewController: UIViewController {
     
     @IBAction func onPlay(_ sender: Any) {
         if (recording != nil) {
-            if (mediaType == Media_Type.Video) {
+            if (mediaType == VIDEO) {
                 if (UserDefaults.standard.bool(forKey: Common.Custom_Player_Key) == true) {
                     guard let urlString = Common.ziggeo?.videos.getVideoUrl(recording!.token) else {
                         return
@@ -161,14 +161,14 @@ class RecordingDetailViewController: UIViewController {
                     present(vc, animated: true, completion: nil)
                     
                 } else {
-                    Common.ziggeo?.playVideo([recording!.token!])
+                    Common.ziggeo?.playVideo(recording!.token!)
                 }
                 
-            } else if (mediaType == Media_Type.Audio) {
-                Common.ziggeo?.startAudioPlayer([recording!.token!])
+            } else if (mediaType == AUDIO) {
+                Common.ziggeo?.startAudioPlayer(recording!.token!)
                 
             } else {
-                Common.ziggeo?.showImage([recording!.token!])
+                Common.ziggeo?.showImage(recording!.token!)
             }
         }
     }